Cleanup error handling to move past problems.
Now all exceptions are output to $stderr and the index is pushed past
the problem so additional parsing can be attempted. I did this to get an
overview for how broken/dumb the current parsing code is. I'm seeing 277
errors and 3135 successfully parsed entries in my current situation, so
- not a bad beginning, but looks like I'll need to dig in a bit more.
Although the current goal is to grab enough of the destination folders
to successfully recreate the original directory structure, so Put Back
actions from Trash can happen successfully.

In other words, I may not need to fully parse the file to get what I
need.