Better handles file_id_headers found in strings
I've largely NOT paid attention to the details @
http://search.cpan.org/~wiml/Mac-Finder-DSStore/DSStoreFormat.pod since
I've come across record types that aren't documented there (I don't
think) and have had a good amount of success percentage-wise in parsing
the file.

Now I'm attempting to work on the edge cases, and with this change I'm
able to correctly parse about half of the ones previously skipped.